backend/nexus_back/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==11.3.0
python-decouple==3.8
psycopg2==2.9.10
dj-database-url==2.1.0